app/backend/requirements.txt
Flask>=2.2
pymongo>=4.0
dnspython>=2.0
flask-jwt-extended>=4.5
bcrypt>=4.0
python-dotenv>=1.0
gridfs
python-magic